feat(payment-service): minor units, transactional outbox, and provider-driven states
Store amounts as BIGINT in the currency's minor units, as the architecture
document specifies. 10000 is USD 100.00. Doing this now costs one migration
over test data; after the ledger, refunds, and settlements exist it would mean
migrating four tables of financial records at once.

Disable Jackson's accept-float-as-int while doing it. Its default is to
truncate 10.99 to 10, which would silently charge the wrong amount. A
fractional amount is now a client error.

Add a transactional outbox. The payment row and its event are written in one
local transaction, so an event can never describe a payment that rolled back
and a payment can never commit without its event. A relay publishes to Kafka
afterwards and only marks a row published once the broker has acknowledged it,
which makes delivery at-least-once rather than best-effort.

Extend the state machine with PENDING_PROVIDER and CANCELLED, and drive it
from inbound events instead of from the client.

Remove POST /api/v1/payments/{id}/status from the public API. With provider
callbacks driving the lifecycle, leaving that route in place would let a
merchant mark its own payment CAPTURED with no acquirer involved. A test pins
that it stays gone.

Transitions applied from events are deliberately tolerant: a redelivery asking
for the state we are already in is a no-op, and an out-of-order one is dropped
rather than thrown. Throwing would put the consumer in a redelivery loop over
a message that can never succeed.

+++ services/payment-service/src/main/java/com/openpay/payment/api/PaymentResponse.java
@@ -1,14 +1,14 @@
 package com.openpay.payment.api;
 
 import com.openpay.payment.domain.PaymentStatus;
-import java.math.BigDecimal;
 import java.time.OffsetDateTime;
 import java.util.UUID;
 
 public record PaymentResponse(
         UUID id,
         PaymentStatus status,
-        BigDecimal amount,
+        /** Amount in the currency's smallest unit. */
+        Long amount,
         String currency,
         OffsetDateTime createdAt,
         OffsetDateTime updatedAt) {
